requirements.txt
# Core Framework
flask==3.1.0
werkzeug==3.1.3

# Database & ORM
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
sqlalchemy==2.0.38
alembic==1.14.1
psycopg2-binary==2.9.10

# API & Documentation
flask-restx>=1.3.0
flask-swagger==0.2.14

# Authentication & Security
flask-jwt-extended==4.6.0
flask-login==0.6.3
flask-security-too==5.4.3
flask-principal==0.4.0
flask-wtf==1.2.1
passlib[bcrypt]==1.7.4
flask-mail>=0.9.1
email-validator>=2.0.0

# Admin Interface
flask-admin==1.6.1
flask-babelex==0.9.4
wtforms==3.1.2

# CORS & Session Management
flask-cors==5.0.1
flask-session>=0.8.0

# Cloud Services
cloudinary==1.42.2

# Google OAuth
google-auth>=2.23.3
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.1

# Production Server
gunicorn==23.0.0

# Configuration & Environment
python-dotenv==1.0.1

# Template Engine
jinja2==3.1.5
markupsafe==3.0.2

# Utilities
click==8.1.8
itsdangerous==2.2.0
certifi>=2025.1.31
urllib3==2.3.0
six==1.17.0
pyyaml==6.0.2
blinker==1.9.0
pyjwt==2.10.1
mako==1.3.9
packaging==24.2
typing-extensions==4.12.2
greenlet==3.1.1
